Verify client version on agent reconnect
Extract the check out of lib/join/joinclient into a new lib/clientversion
package and run it at the top of newClient, before any transport is
attempted. It stays best-effort, failing open when the cluster's
advertised minimum can't be fetched or parsed, and is bypassed
with a warning when --skip-version-check is set.
